macos cpp11

+++ muduo/net/poller/PollPoller.cc
@@ -29,7 +29,7 @@ PollPoller::~PollPoller() = default;
 Timestamp PollPoller::poll(int timeoutMs, ChannelList* activeChannels)
 {
   // XXX pollfds_ shouldn't change
-  int numEvents = ::poll(&*pollfds_.begin(), pollfds_.size(), timeoutMs);
+  int numEvents = ::poll(&*pollfds_.begin(), (unsigned int)(pollfds_.size()), timeoutMs);
   int savedErrno = errno;
   Timestamp now(Timestamp::now());
   if (numEvents > 0)
